+++ core/src/main/scala/helper/Helper.scala
@@ -22,6 +22,13 @@ import java.security.cert.X509Certificate
 
 import config.AppConfig
 import javax.net.ssl.{HostnameVerifier, SSLSession, X509TrustManager}
+import org.apache.kafka.clients.CommonClientConfigs
+import org.apache.kafka.common.config.SslConfigs
+import org.apache.spark.sql.Column
+import org.apache.spark.sql.avro.SchemaConverters
+import za.co.absa.abris.avro.read.confluent.SchemaManagerFactory
+import za.co.absa.abris.avro.registry.SchemaSubject
+import za.co.absa.abris.config.{AbrisConfig, ToAvroConfig, ToSchemaDownloadingConfigFragment, ToStrategyConfigFragment}
 
 class Helper(appConfig: AppConfig) {
 
@@ -124,21 +131,21 @@ class Helper(appConfig: AppConfig) {
   }
 
   /**
-   * Returns the text (content) from a REST URL as a String.
-   *
-   * @param url            The full URL to connect to.
-   * @param connectTimeout Sets a specified timeout value, in milliseconds,
-   *                       to be used when opening a communications link to the resource referenced
-   *                       by this URLConnection. If the timeout expires before the connection can
-   *                       be established, a java.net.SocketTimeoutException
-   *                       is raised. A timeout of zero is interpreted as an infinite timeout.
-   *                       Defaults to 10000 ms.
-   * @param readTimeout    If the timeout expires before there is data available
-   *                       for read, a java.net.SocketTimeoutException is raised. A timeout of zero
-   *                       is interpreted as an infinite timeout. Defaults to 10000 ms.
-   * @param requestMethod  Defaults to "GET". (Other methods have not been tested.)
-   *
-   */
+    * Returns the text (content) from a REST URL as a String.
+    *
+    * @param url            The full URL to connect to.
+    * @param connectTimeout Sets a specified timeout value, in milliseconds,
+    *                       to be used when opening a communications link to the resource referenced
+    *                       by this URLConnection. If the timeout expires before the connection can
+    *                       be established, a java.net.SocketTimeoutException
+    *                       is raised. A timeout of zero is interpreted as an infinite timeout.
+    *                       Defaults to 10000 ms.
+    * @param readTimeout    If the timeout expires before there is data available
+    *                       for read, a java.net.SocketTimeoutException is raised. A timeout of zero
+    *                       is interpreted as an infinite timeout. Defaults to 10000 ms.
+    * @param requestMethod  Defaults to "GET". (Other methods have not been tested.)
+    *
+    */
   @throws(classOf[java.io.IOException])
   @throws(classOf[java.net.SocketTimeoutException])
   def get(url: String,
@@ -198,9 +205,71 @@ class Helper(appConfig: AppConfig) {
     }
   }
 
-  def buildMongoURI(login: String, password: String, cluster: String, replicaSet: String, autheticationDatabase: String, database: String, collection: String, authenticationEnabled: Boolean): String = {
+  def buildSecureKafkaProperties(bootstrapServers: String,
+                                 schemaRegistries: String,
+                                 keyStorePath: String,
+                                 trustStorePath: String,
+                                 keyStorePassword: String,
+                                 trustStorePassword: String,
+                                 keyPassword: String): Map[String, String] = {
+
+    var props = Map("bootstrap.servers" -> bootstrapServers, "schema.registry.url" -> schemaRegistries)
+
+    if (keyStorePath != "null" || trustStorePath != "null") {
+      props += (CommonClientConfigs.SECURITY_PROTOCOL_CONFIG -> "SSL")
+      if (keyStorePath != null)
+        props += (SslConfigs.SSL_KEYSTORE_LOCATION_CONFIG -> keyStorePath)
+      if (trustStorePath != null)
+        props += (SslConfigs.SSL_TRUSTSTORE_LOCATION_CONFIG -> trustStorePath)
+      if (keyStorePassword != null)
+        props += (SslConfigs.SSL_KEYSTORE_PASSWORD_CONFIG -> keyStorePassword)
+      if (trustStorePassword != null)
+        props += (SslConfigs.SSL_TRUSTSTORE_PASSWORD_CONFIG -> trustStorePassword)
+      if (keyPassword != null)
+        props += (SslConfigs.SSL_KEY_PASSWORD_CONFIG -> keyPassword)
+      props += (SslConfigs.SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG -> "")
+    }
+
+    props
+
+  }
+
+  def GetToAvroConfig(topic: String, schemaRegistryUrl: String, dfColumn: Column, schemaVersion: Option[Int] = None, isKey: Boolean = false, subjectNamingStrategy: String = "TopicNameStrategy" /*other options are RecordNameStrategy, TopicRecordNameStrategy*/ , subjectRecordName: Option[String] = None, subjectRecordNamespace: Option[String] = None): ToAvroConfig = {
+    //get the specified schema version
+    //if not specified, then get the latest schema from Schema Registry
+    //if the topic does not have a schema then create and register the schema
+    //applies to both key and value
+    val subject = if (subjectNamingStrategy.equalsIgnoreCase("TopicRecordNameStrategy")) SchemaSubject.usingTopicRecordNameStrategy(topicName = topic, recordName = subjectRecordName.getOrElse(""), recordNamespace = subjectRecordNamespace.getOrElse("")) else if (subjectNamingStrategy.equalsIgnoreCase("RecordNameStrategy")) SchemaSubject.usingRecordNameStrategy(recordName = subjectRecordName.getOrElse(""), recordNamespace = subjectRecordNamespace.getOrElse("")) else SchemaSubject.usingTopicNameStrategy(topicName = topic, isKey = isKey) // Use isKey=true for the key schema and isKey=false for the value schema
+    val schemaRegistryClientConfig = Map(AbrisConfig.SCHEMA_REGISTRY_URL -> schemaRegistryUrl)
+    val schemaManager = SchemaManagerFactory.create(schemaRegistryClientConfig)
+    val expression = dfColumn.expr
+    val dataSchema = SchemaConverters.toAvroType(expression.dataType, expression.nullable)
+    println((if (isKey) "key" else "value") + " subject = " + subject.asString)
+    println((if (isKey) "key" else "value") + " avro schema inferred from data  = " + dataSchema.toString())
+    var toAvroConfig: ToAvroConfig = null
+    if (schemaManager.exists(subject)) {
+      val avroConfigFragment = AbrisConfig
+        .toConfluentAvro
+      val toStrategyConfigFragment: ToStrategyConfigFragment = if (schemaVersion.isEmpty) avroConfigFragment.downloadSchemaByLatestVersion
+      else avroConfigFragment.downloadSchemaByVersion(schemaVersion.get)
+      val schemaDownloadingConfigFragment: ToSchemaDownloadingConfigFragment = if (subjectNamingStrategy.equalsIgnoreCase("TopicRecordNameStrategy")) toStrategyConfigFragment.andTopicRecordNameStrategy(topicName = topic, recordName = subjectRecordName.getOrElse(""), recordNamespace = subjectRecordNamespace.getOrElse("")) else if (subjectNamingStrategy.equalsIgnoreCase("RecordNameStrategy")) toStrategyConfigFragment.andRecordNameStrategy(recordName = subjectRecordName.getOrElse(""), recordNamespace = subjectRecordNamespace.getOrElse("")) else toStrategyConfigFragment.andTopicNameStrategy(topic, isKey = isKey)
+      toAvroConfig = schemaDownloadingConfigFragment
+        .usingSchemaRegistry(schemaRegistryUrl)
+    }
+    else {
+      val schemaId = schemaManager.register(subject, dataSchema)
+      toAvroConfig = AbrisConfig
+        .toConfluentAvro
+        .downloadSchemaById(schemaId)
+        .usingSchemaRegistry(schemaRegistryUrl)
+    }
+    println((if (isKey) "key" else "value") + " avro schema expected by schema registry  = " + toAvroConfig.schemaString)
+    toAvroConfig
+  }
+
+  def buildMongoURI(login: String, password: String, cluster: String, replicaSet: String, autheticationDatabase: String, database: String, collection: String, authenticationEnabled: Boolean, sslEnabled: String): String = {
     if (authenticationEnabled) {
-      "mongodb://" + URLEncoder.encode(login, "UTF-8") + ":" + URLEncoder.encode(password, "UTF-8") + "@" + cluster + ":27017/" + database + "." + collection + "?authSource=" + (if (autheticationDatabase != "") autheticationDatabase else "admin") + (if (replicaSet == null) "" else "&replicaSet=" + replicaSet)
+      "mongodb://" + URLEncoder.encode(login, "UTF-8") + ":" + URLEncoder.encode(password, "UTF-8") + "@" + cluster + ":27017/" + database + "." + collection + "?authSource=" + (if (autheticationDatabase != "") autheticationDatabase else "admin") + (if (replicaSet == null) "" else "&replicaSet=" + replicaSet) + (if (sslEnabled == "true") "&ssl=true&sslInvalidHostNameAllowed=true" else "")
     } else {
       "mongodb://" + cluster + ":27017/" + database + "." + collection + (if (replicaSet == null) "" else "&replicaSet=" + replicaSet)
     }