+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.1.19] - 2021-02-12
+### Changed
+- api/src/main/java/com/homeaway/datapullclient/config/EMRProperties.java: Upgraded EMR version to 5.31.0
+- core/docker_spark_server/Dockerfile: Upgraded Spark version to 2.4.6, downgraded Hadoop version to 2.10.0
+- core/src/main/resources/Samples/Input_Json_Specification.json: Upgraded Kafka destination spec to match abris 4.0.1 capabilities
+- core/src/main/resources/Samples/Input_Sample_Join_Heterogeneous_Sources.json: Fixed file spelling
+- core/src/main/resources/Samples/Input_Sample_s3_to_kafka.json: Upgraded example to match abris 4.0.1 capabilities
+- core/src/main/scala/core/DataFramFromTo.scala: Removed unnecessary println's, Use s3a when running locally, s3 otherwise, upgraded ABRiS to 4.0.1
+- core/src/main/scala/core/DataPull.scala: Use s3a when running locally, s3 otherwise
+- core/src/main/scala/core/Migration.scala: Removed unnecessary println's, upgraded ABRiS to 4.0.1
+- pom.xml: Upgraded Spark version to 2.4.6, downgraded Hadoop version to 2.10.0, removed shaded jar, upgraded ABRiS to 4.0.1
+- README.md: Upgraded Spark version to 2.4.6, downgraded Hadoop version to 2.10.0
+
+
 ## [0.2.1] - 2021-01-26
 ### Changed
 api/src/main/java/com/homeaway/datapullclient/process/DataPullTask.java: S3 Cross-Account Access Support.
@@ -15,6 +29,7 @@ api/src/main/java/com/homeaway/datapullclient/process/DataPullTask.java: S3 Cros
 - Removed unnecessary environment variables
 ### Removed
 - Removed the MS DOS batch script for ECS Deploy
+
 
 ## [0.1.18] - 2020-11-05
 ### Changed